+++ src/main/java/org/lamisplus/modules/pharmacy/repository/DrugOrderRepository.java
@@ -25,5 +25,5 @@ public interface DrugOrderRepository extends JpaRepository<DrugOrder, Long> {
     @Query(value = "SELECT * FROM drug_order WHERE patient_id=?1 GROUP BY prescription_group_id, id ORDER BY id DESC", nativeQuery = true)
     List<DrugOrder> findAllByPatientIdGroupByPrescriptionGroupIdOrderById(Long patientId);
 
-    List<DrugOrder> findAllByArchivedAndVisitId(int archived, Integer visitId);
+    List<DrugOrder> findAllByArchivedAndVisitId(int archived, Long visitId);
 }
